+++ lua/plugins/config/smart_pairs.lua
@@ -1,9 +1,0 @@
-return function(use)
-	use({
-		'ZhiyuanLck/smart-pairs',
-		event = 'InsertEnter',
-		config = function()
-			require('pairs'):setup()
-		end,
-	})
-end